requirements.txt
pandas
faker
numpy
streamlit
sqlite-utils
python-dotenv
langchain
langchain-groq
langgraph
plotly

